Move History struct to Types.swift, minor changes
