--- SlackKit/Sources/History.swift
@@ -1,49 +1,0 @@
-//
-// Attachment.swift
-//
-// Copyright © 2016 Peter Zignego. All rights reserved.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-import Foundation
-
-public struct History {
-    public let ok: Bool?
-    public let latest: NSDate?
-    public let messages: [Message]?
-    public let hasMore: Bool?
-
-    internal init?(history: [String: AnyObject]?) {
-        ok = history?["ok"] as? Bool
-        if let latestStr = history?["latest"] as? String, latestDouble = Double(latestStr) {
-            latest = NSDate(timeIntervalSince1970: NSTimeInterval(latestDouble))
-        } else {
-            latest = nil
-        }
-        messages = [Message]()
-        if let ms = history?["messages"] as? [[String: AnyObject]] {
-            for m in ms {
-                if let message = Message(message: m) {
-                    messages?.append(message)
-                }
-            }
-        }
-        hasMore = history?["has_more"] as? Bool
-    }
-}
